update_nrpe_checks(): fetch nagios_hostname from nrpe-external-master relation
This is necessary to emit valid nagios configs when related to
n-e-m that has nagios_hostname_type=host, which can be the case
with e.g. MAAS deployments.

Fall back to our previous behavour of constructing the hostname
ourselves if nagios_hostname is not set or if there are multiple
relations found.