Revert change to distribute client credentials using leadership
This reverts commits 926a86e9411f3d7d5992b14f35d520f502cd3bff & 14847db0ac37ddc735118ef2dc65a0b07736375f

This change allowed a subordinate to connect to PostgreSQL using a
container scoped relation, which is a dubious use case. However, it also
introduced race conditions where client credentials were distributed to
clients before the matching resources had been created on the master.
Fixing this would make the charm more fragile, as the leader would need
to create resources remotely on the master and this would fail if the
master was not available at that time.