Switchover action
The new switchover action allows the operator to switch to a new master
of their choosing. As the existing master is still live, we can
first setup a replication loop to ensure all transactions are
replicated before promotion. This should allow it to be used even
on busy systems, without data loss.